pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sinchan-env"
version = "0.1.0"
description = "ShinChan Life Simulator — An OpenEnv RL environment where LLMs learn from Shin-chan's mistakes!"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "fastmcp>=0.1.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=4.0.0",
]
training = [
    "trl>=0.12.0",
    "datasets>=2.14.0",
    "torch>=2.0.0",
    "transformers>=5.2.0",
    "jmespath>=1.0.0",
    "peft>=0.11.0",
    "bitsandbytes>=0.42.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "sinchan_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sinchan_env", "sinchan_env.server"]
package-dir = { "sinchan_env" = ".", "sinchan_env.server" = "server" }

[tool.setuptools.package-data]
"sinchan_env.server" = ["static/sinchan/index.html", "static/blog.md"]
